# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Food Delivery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # CATEGORIES
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items for a restaurant
      parameters:
        - name: restaurantId
          in: query
          required: true
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CART
...
  /cart/add:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/remove/{itemId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin list all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/restaurants:
    get:
      operationId: adminListRestaurants
      tags: [admin]
      summary: Admin list all restaurants
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Restaurants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantPage"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, restaurant_owner, delivery_partner, admin]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, restaurant_owner, delivery_partner, admin]
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string

    # CATEGORIES
...
    Restaurant:
      type: object
      required: [id, ownerId, name, address, status]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, pending]
        avgRating:
          type: ["number", "null"]
        totalReviews:
          type: ["integer", "null"]
...
    RestaurantPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Restaurant"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    MenuItem:
      type: object
      required: [id, restaurantId, name, price, available]
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        image:
          type: ["string", "null"]
        available:
          type: boolean
    MenuItemInput:
      type: object
      required: [restaurantId, name, price]
      properties:
        restaurantId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        image:
          type: ["string", "null"]
        available:
          type: boolean
    MenuItemUpdate:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        image:
          type: ["string", "null"]
        available:
          type: boolean

    # CART
    Cart:
      type: object
      required: [id, userId, items, total]
      properties:
        id:
          type: integer
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        restaurantId:
          type: ["integer", "null"]
        restaurantName:
          type: ["string", "null"]
    CartItem:
      type: object
      required: [id, menuItemId, name, price, quantity, subtotal]
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
        image:
          type: ["string", "null"]
    CartItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer

    # ORDERS
...
    UserPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, restaurant_owner, delivery_partner, admin]
        active:
          type: boolean